Use int32_t in place for long, to avoid LP64/LLP64 issue.

+++ src/common/sequence.cpp
@@ -91,7 +91,7 @@ uint32_t CALLBACK Sequencer::ThreadEntry(void* arg) {
 //  length  実行する時間 (0.01ms)
 //  eff     実効クロック
 //
-inline void Sequencer::Execute(long clk, long length, long eff) {
+inline void Sequencer::Execute(int32_t clk, int32_t length, int32_t eff) {
   CriticalSection::Lock lock(cs);
   execcount += clk * vm->Proceed(length, clk, eff);
 }
@@ -155,7 +155,7 @@ void Sequencer::ExecuteAsynchronus() {
 // ---------------------------------------------------------------------------
 //  実行クロックカウントの値を返し、カウンタをリセット
 //
-long Sequencer::GetExecCount() {
+int32_t Sequencer::GetExecCount() {
   //  CriticalSection::Lock lock(cs); // 正確な値が必要なときは有効にする
 
   int i = execcount;
